move ModelFairnessMetric to another file
